Refactor top methods
We can use slice to make an array with 3 items from the sorted array
(starting at index 0).

#map lets us create a new array by iterating over each of the items in
the array and returning  a specific value for each one. In this case,
we’re making an array with all of the letters/words.

Functionality wise, this does the same thing you were doing. Just
thought it was a good place to show how slice and map could be used to
make this more concise.